Rename rpoject and package to flask-storage-helpers to avoid conflict with already existing Flask-Storage on Pypi
