+++ README.rst
@@ -1,4 +1,4 @@
-flask-storage
-=============
+flask-storage-helpers
+=====================
 
 Various file storage backends for Flask applications.
